Allow requesting _type and __typename on hash renderer
The former is a somewhat "nice to have" in either case, the latter
is specific to graphql. Since these are opt-in with sparse fieldsets,
I see no reason to complicate things with a GQL check here - if you
want it you got it.

The `Util::Class` method will be re-used by `graphiti_graphql`, there's
a little bit of chocolate mixing with our peanut butter here but I'm
OK with it.