Ensure remote resources with with vanilla json renderer
We do this absolutely wild hack of setting the resource to `1` when
serializing a remote resource, to overcome an issue in jsonapi-rb.
Because the hash renderer is doing a bit more than it used to, add a
test to ensure something like `@resource.polymorphic?` won't break
things when `@resource` is `1`.