Added filter_group DSL:
Added filter_group DSL to require filtering by all or any of the
filters specified in the filter group. For example if we want to
filter by either employee_id or employee_uid field.

This commit raises an error if any/all of the required filters
are not passed into the request.

filter_group: [:filter1, filter1], required: :all | :any