Add validation error when data/type missing
Addresses https://github.com/graphiti-api/graphiti/issues/339

A few other things came up - one is that all requests, even reads, go
through the validator. Ideally the check is further upstream but I
tried to change the least amount of code while making this explicit.

The other is that I'm normalizing the params in the constructor, so
we can always deal with something normalized instead of worrying about
raw vs normalized. This has a downstream effect of changing the error
message for a Rails4-specific test you can see in the diff (if
anything, probably more correct now).