Allow unwritable PK when associating on create
Imagine updating a `Position`, changing the `Employee` it is associated
with. Since we're justassociating, we don't need `Employee`s PK to be
writable. But because our internals end up saving the `Employee`
regardless - we just walk the tree, and `id` is part of the attributes
considered - so we'd normally validate the PK, even though we really
don't need to.

We already avoided validating the PK when *updating* the `Position`,
this PR also considers *creating* the `Position` and associating to a
pre-existing `Employee`. We do this by looking at the `method` of
each sidepost, instad of the overall action being called.