Support page[offset] param
Say you want to start paging, 10 per page, but skip the first 8 results.
Previously we didn't support this since the page count isn't a multiple
of the offset - now we can! This will serve not only as a useful param
in-and-of-itself, but also serve as the basis for future cursor
pagination, which will default to offset-based (as opposed to
stable-id-based) by default.

Small notes on backwards compatibility. If you have a custom `paginate`
override we can't change the existing order so it's:

```ruby
paginate do |scope, current_page, per_page, context, offset|
```

Instead of `offset, context`.

Also, we check the adapter arity to see if we should pass `offset` -
this should accomodate existing non-AR adapters.