Support schema for filters of extra attributes
Fixes https://github.com/graphiti-api/graphiti/issues/378

Defining a filter for an extra attribute causes the schema generation to
fail since it doesn't consider extra attributes.

This change allows filters (and sorts) of extra attributes to be
included in the generated schema. It also adds support for the `schema`
attribute option for extra attributes for parity with attributes.