Refactor future using methods into their own methods
This makes it clearer what's going on when a method is called by
enforcing that either the method will return a future, or a resolved
value. It simplifies the code by removing the conditional statements
that were needed to check whether a future or a resolved value was being
returned, which was necessary to know whether the code was being called
async or not.

It also ensures that sync code always has a synchronous entry point to
call that will always block.